feat: Implement modular engine, camera system, and custom scripts
This commit introduces a major architectural refactor to the game engine, making it more modular, extensible, and data-driven.

The key changes include:

- **Modular Engine:** The monolithic `engine.js` has been broken down into smaller, more focused modules for scene loading, script management, and the game loop. These are now located in the `js/engine/` directory.

- **Flexible Camera System:** The camera is now defined in `game.json`, allowing for easy configuration of its properties. A new `camera-follow` script has been added to demonstrate dynamic camera behavior.

- **Custom, Data-Driven Scripts:** Game object behaviors are no longer hardcoded. Instead, they are defined as individual script modules in the `js/scripts/` directory and attached to objects in `game.json`. This allows for creating new behaviors without modifying the core engine.

- **Documentation:** A `README.md` file has been added to explain the new architecture and provide instructions on how to use the new features.
