+++ game.json
@@ -2,6 +2,24 @@
   "scene": {
     "background": "0xaaaaaa"
   },
+  "camera": {
+    "type": "PerspectiveCamera",
+    "fov": 75,
+    "near": 0.1,
+    "far": 15,
+    "position": {
+      "x": 0,
+      "y": 5,
+      "z": 5
+    },
+    "scripts": [
+      {
+        "type": "camera-follow",
+        "target": "main_cube",
+        "smooth_speed": 0.125
+      }
+    ]
+  },
   "objects": [
     {
       "name": "main_cube",
@@ -26,8 +44,7 @@
           "type": "gravity"
         },
         {
-          "type": "keyboardInput",
-          "speed": 5
+          "type": "keyboard-input"
         }
       ]
     },
