+++ index.html
@@ -18,6 +18,6 @@
 				}
 			}
 		</script>
-		<script type="module" src="js/engine.js"></script>
+		<script type="module" src="js/engine/main.js"></script>
 	</body>
 </html>
