+++ js/engine/main.js
@@ -1,0 +1,39 @@
+import * as THREE from 'three';
+import { loadScene } from './scene-loader.js';
+import { createGameLoop } from './game-loop.js';
+
+async function main() {
+    const canvas = document.querySelector('#c');
+    const renderer = new THREE.WebGLRenderer({ antialias: true, canvas });
+
+    const { scene, camera, gameObjects } = await loadScene();
+
+    const gameLoop = createGameLoop(renderer, scene, camera, gameObjects);
+
+    function resizeRendererToDisplaySize(renderer) {
+        const canvas = renderer.domElement;
+        const width = canvas.clientWidth;
+        const height = canvas.clientHeight;
+        const needResize = canvas.width !== width || canvas.height !== height;
+        if (needResize) {
+            renderer.setSize(width, height, false);
+        }
+        return needResize;
+    }
+
+    function render() {
+        if (resizeRendererToDisplaySize(renderer)) {
+            const canvas = renderer.domElement;
+            camera.aspect = canvas.clientWidth / canvas.clientHeight;
+            camera.updateProjectionMatrix();
+        }
+
+        gameLoop(performance.now());
+        renderer.render(scene, camera);
+        requestAnimationFrame(render);
+    }
+
+    requestAnimationFrame(render);
+}
+
+main();
